third_party/WebKit/LayoutTests/http/tests/devtools/indexeddb/transaction-promise-console-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Ensure transactions created within Promise callbacks are not deactivated due to console activity

PASS: Transaction is still active
